+++ packages/services/f-analytics/CHANGELOG.md
@@ -3,6 +3,17 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v0.8.0
+------------------------------
+*August 9, 2021*
+
+### Changed
+- Added push ad-hoc event method to global var
+- Moved the preparation of the server side plaform data to a plugin.
+- Refactored the Store to allow store serverside events to be stored until clientside.
+
+
 v0.5.0
 ------------------------------
 *July 28, 2021*